requirements.txt
fastapi==0.135.1
openai==2.26.0
pydantic==2.12.5
python-dotenv==1.2.2
Requests==2.32.5
SPARQLWrapper==2.0.0
